requirements.txt
# RFID Access Control System Dependencies
# Versione ottimizzata per Raspberry Pi

# Core RFID Support - Versione compatibile
mfrc522==0.0.7

# Alternativa più stabile (se la precedente da problemi)
# pip-autoremove mfrc522
# pip install pi-rc522

# GPIO Control (Raspberry Pi)
RPi.GPIO==0.7.1

# MQTT Communication
paho-mqtt==1.6.1

# Configuration Management
python-dotenv==1.0.0
